[v1] - Building Client Console

+++ src/CalcOperations.Starship.Business/Entities/StopCalculationResult.cs
@@ -3,6 +3,6 @@ namespace CalcOperations.Starship.Business.Entities
     public class StopCalculationResult
     {
         public string StarshipName { get; set; }
-        public int TotalStops { get; set; }
+        public double TotalStops { get; set; }
     }
 }
